requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
openai==1.3.5
python-pptx==0.6.23
python-dotenv==1.0.0